Added grid layout
Uses a overlaying grid layout for cordinates.
Where cordinate start at a given latlon position in top left overner of the map and moves downward towards the right in specified x and y grid size.
And makes a request to google places api to get all restaurants in a given radius and adds them to db.